Bump minimum required requests lib version to 2.4.3
The max_retries HTTPAdapter init argument was not added until requests
v1.2.0 so if we're going to use that initializer then we should bump the
minimum requests version from 1.0.0 to 1.2.3 (the last patch on 1.2).
Since we're doing a major revision update anyway, we've decided to also
move the requests dependency a bit further along. Unfortunately requests
v2.0-2.3 are not quite functional, so here we are at v2.4(.3).